requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pydantic>=2.7.1
pydantic-settings>=2.3.0
google-genai>=1.5.0
google-cloud-texttospeech>=2.16.0
google-cloud-speech>=2.21.0
httpx>=0.27.0
python-multipart>=0.0.9
websockets>=12.0
alembic>=1.13.1
requests>=2.31.0